+++ perf/tests/test_timeit.py
@@ -7,6 +7,7 @@ import unittest
 
 import perf
 from perf import tests
+from perf.text_runner import _which as which
 
 
 SLEEP = 'time.sleep(1e-3)'
@@ -169,6 +170,39 @@ class TestTimeit(unittest.TestCase):
 
         self.assertIn('Traceback (most recent call last):', stderr)
         self.assertIn("NameError", stderr)
+
+    def check_python_option(self, python):
+        path = os.path.realpath(os.path.dirname(perf.__file__))
+        env = dict(os.environ)
+        env['PYTHONPATH'] = path
+
+        abs_python = which(python, path=os.defpath)
+        if not abs_python:
+            self.skipTest('missing Python executable: %s' % python)
+        python = os.path.realpath(abs_python)
+
+        # Run benchmark to check if --python works
+        args = [sys.executable,
+                '-m', 'perf', 'timeit',
+                '--metadata', '--debug-single-sample',
+                '--python', python,
+                '-s', 'import time',
+                SLEEP]
+        proc = subprocess.Popen(args,
+                                stdout=subprocess.PIPE,
+                                stderr=subprocess.PIPE,
+                                universal_newlines=True,
+                                env=env)
+        stdout, stderr = proc.communicate()
+        self.assertEqual(proc.returncode, 0, stderr)
+
+        self.assertIn("python_executable: %s" % python, stdout)
+
+    def test_python_option_py2(self):
+        self.check_python_option('python2')
+
+    def test_python_option_py3(self):
+        self.check_python_option('python3')
 
 
 if __name__ == "__main__":