+++ perf/text_runner.py
@@ -23,6 +23,74 @@ except ImportError:
     psutil = None
 
 
+try:
+    # Python 3.3
+    from shutil import which as _which
+except ImportError:
+    # Backport shutil.which() from Python 3.6
+    def _which(cmd, mode=os.F_OK | os.X_OK, path=None):
+        """Given a command, mode, and a PATH string, return the path which
+        conforms to the given mode on the PATH, or None if there is no such
+        file.
+
+        `mode` defaults to os.F_OK | os.X_OK. `path` defaults to the result
+        of os.environ.get("PATH"), or can be overridden with a custom search
+        path.
+
+        """
+        # Check that a given file can be accessed with the correct mode.
+        # Additionally check that `file` is not a directory, as on Windows
+        # directories pass the os.access check.
+        def _access_check(fn, mode):
+            return (os.path.exists(fn) and os.access(fn, mode)
+                    and not os.path.isdir(fn))
+
+        # If we're given a path with a directory part, look it up directly rather
+        # than referring to PATH directories. This includes checking relative to the
+        # current directory, e.g. ./script
+        if os.path.dirname(cmd):
+            if _access_check(cmd, mode):
+                return cmd
+            return None
+
+        if path is None:
+            path = os.environ.get("PATH", os.defpath)
+        if not path:
+            return None
+        path = path.split(os.pathsep)
+
+        if sys.platform == "win32":
+            # The current directory takes precedence on Windows.
+            if os.curdir not in path:
+                path.insert(0, os.curdir)
+
+            # PATHEXT is necessary to check on Windows.
+            pathext = os.environ.get("PATHEXT", "").split(os.pathsep)
+            # See if the given file matches any of the expected path extensions.
+            # This will allow us to short circuit when given "python.exe".
+            # If it does match, only test that one, otherwise we have to try
+            # others.
+            if any(cmd.lower().endswith(ext.lower()) for ext in pathext):
+                files = [cmd]
+            else:
+                files = [cmd + ext for ext in pathext]
+        else:
+            # On other platforms you don't have things like PATHEXT to tell you
+            # what file suffixes are executable, so just pass on cmd as-is.
+            files = [cmd]
+
+        seen = set()
+        for dir in path:
+            normdir = os.path.normcase(dir)
+            if normdir not in seen:
+                seen.add(normdir)
+                for thefile in files:
+                    name = os.path.join(dir, thefile)
+                    if _access_check(name, mode):
+                        return name
+        return None
+
+
 def _run_cmd(args, env):
     proc = subprocess.Popen(args,
                             universal_newlines=True,
@@ -103,6 +171,9 @@ class TextRunner:
         # (sys.executable, sys.argv[0]) by default. For example,
         # "python3 -m perf timeit" sets program_args to
         # (sys.executable, '-m', 'perf', 'timeit').
+        #
+        # The first item is overriden by the value of the --python command line
+        # option.
         self.program_args = (sys.executable, sys.argv[0])
 
         # Number of inner-loops of the sample_func for bench_sample_func()
@@ -191,6 +262,10 @@ class TextRunner:
                             help='Comma-separated list of environment '
                                  'variables inherited by worker child '
                                  'processes.')
+        parser.add_argument("--python", default=sys.executable,
+                            help='Python executable '
+                                 '(default: use running Python, '
+                                 'sys.executable)')
 
         memory = parser.add_mutually_exclusive_group()
         memory.add_argument('--tracemalloc', action="store_true",
@@ -245,6 +320,16 @@ class TextRunner:
                 print("ERROR: unable to track the memory usage "
                       "(--track-memory): %s" % err_msg)
                 sys.exit(1)
+
+        # Replace "~" with the user home directory
+        args.python = os.path.expanduser(args.python)
+        # Try to the absolute path to the binary
+        abs_python = _which(args.python)
+        if not abs_python:
+            print("ERROR: Unable to locate the Python executable: %r"
+                  % args.python)
+            sys.exit(1)
+        args.python = os.path.realpath(abs_python)
 
     def parse_args(self, args=None):
         if self.args is None:
@@ -545,6 +630,7 @@ class TextRunner:
 
         cmd = []
         cmd.extend(self.program_args)
+        cmd[0] = args.python
         cmd.extend(('--worker', '--stdout',
                     '--samples', str(args.samples),
                     '--warmups', str(args.warmups),